Merged PR 69: UI Updates and Warning Fixes
This PR contains UI updates for the following:
- Login Landing Page
- Signout Landing Page
- Footer

It also contains a fix for an issue for nested `<a>`s. Created the fix and updated the `<a>` tags used throughout the project with `<Link>` components for uniformity.

Finally, some docstrings were and code refactors were made to a couple of the `_services/` functions.

Related work items: #1048